Example like ECDSA.recover from OpenZeppelin
This example shows how to get the equivalent of OpenZeppelin ECDSA.recover for long 65-byte Ethereum signatures
(https://docs.openzeppelin.com/contracts/2.x/api/cryptography#ECDSA-recover-bytes32-bytes-)
Short 64-byte Ethereum signatures require some changes to the code.
This can be used in applications requiring some compatibility with Ethereum.
For pure Algorand applications, ed25519 is strongly recommended instead of ecdsa.

Like the other examples, this example has *NOT* been audited and *MUST NOT* be used in production code.
If this example is used in production code, a full audit is necessary due to the nature of this example
(use of cryptography).

Changes:
* adding the example
* updating requirements.md to use a released version of pyteal
  (now that this is possible)
* updating the README to indicate that sandbox can be in
  release mode now